Norwegian translations of the Acquis Communautaire
These are Norwegian translations of the European acts and regulations known as Acquis Communautaire. In connection with the META-NORD project, prof. Koenraad De Smedt at UiB obtained access to the collection from the Norwegian EEA Secretariat, a part of the EEA/EFTA Section of UD. The EEA Secretariat has given UiB permission to incorporate these texts in the JRC-Acquis Multilingual Parallel Corpus.